+++ service/controller/managementcluster/resource.go
@@ -260,6 +260,7 @@ func newResources(config resourcesConfig) ([]resource.Interface, error) {
 
 			AdditionalScrapeConfigs: config.AdditionalScrapeConfigs,
 			Bastions:                config.Bastions,
+			Customer:                config.Customer,
 			Provider:                config.Provider,
 			Installation:            config.Installation,
 			Mayu:                    config.Mayu,