+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add Cluster Service Priority label.
+- Add customer and organization label to metrics.
 
 ## [3.4.3] - 2022-05-10